Optimized IBE & WIBE encryption schemes
Due to some design issues, the original IBE & WIBE classes had to first serialize a String instance, which was immediately deserialized during the encryption process. For the decryption process, the same happened but in the opposite direction. This issue has now been fixed, which improves the performance of both algorithms.