Security issue resolved.
The IBE segments of the attestations were originally encrypted using the storage layer identifiers for the namespace attestations of the receivers as IBE identifiers. However, the IBE PKGS are included in the proof information segments for the provers, which allows them to decrypt attestations as long as they know the IBE identifiers. Since these storage layer identifiers are ALWAYS included in the attestations in the plaintext header, this would allow the provers to decrypt any attestation issued by an issuer, as long as the issuer issued an attestation which is part of his proof.
This issue is now fixed. The IBE segments are encrypted using the RTree policy for the namespace attestation of the owner of the resources. This only allows the prover to decrypt attestations which can actually be used for his proof.